+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-09 11:15:11
-// Commit: fix(core): fix memory leak by cancelling core subscription
+// Updated at: 2026-04-09 14:08:46
+// Commit: test(core): implement widget integration tests for core
 
 class coreComponentState {
-  final String commitRef = '213';
-  final String lastMessage = 'fix(core): fix memory leak by cancelling core subscription';
-  final DateTime timestamp = DateTime.parse('2026-04-09T11:15:11+05:00');
+  final String commitRef = '225';
+  final String lastMessage = 'test(core): implement widget integration tests for core';
+  final DateTime timestamp = DateTime.parse('2026-04-09T14:08:46+05:00');
 }